refactor: #161 <Booking> 코드 리팩토링
