--- com.taken_seat.booking-service/src/main/java/com/taken_seat/booking_service/booking/application/dto/command/BookingPageReadCommand.java
@@ -1,13 +1,0 @@
-package com.taken_seat.booking_service.booking.application.dto.command;
-
-import java.util.UUID;
-
-import org.springframework.data.domain.Pageable;
-
-public record BookingPageReadCommand(
-	UUID userId,
-	String email,
-	String role,
-	Pageable pageable
-) {
-}
