+++ com.taken_seat.booking-service/src/main/java/com/taken_seat/booking_service/booking/application/dto/query/BookingStatusQuery.java
@@ -1,0 +1,9 @@
+package com.taken_seat.booking_service.booking.application.dto.query;
+
+import java.util.UUID;
+
+public record BookingStatusQuery(
+	UUID userId,
+	UUID performanceId
+) {
+}
