+++ com.taken_seat.booking-service/src/main/java/com/taken_seat/booking_service/booking/application/service/BookingClientService.java
@@ -5,13 +5,10 @@ import java.util.UUID;
 import com.taken_seat.booking_service.booking.presentation.dto.response.SeatLayoutResponseDto;
 import com.taken_seat.common_service.dto.request.BookingSeatClientRequestDto;
 import com.taken_seat.common_service.dto.response.BookingSeatClientResponseDto;
-import com.taken_seat.common_service.dto.response.BookingStatusDto;
 import com.taken_seat.common_service.dto.response.PerformanceStartTimeDto;
 import com.taken_seat.common_service.dto.response.TicketPerformanceClientResponse;
 
 public interface BookingClientService {
-	BookingStatusDto getBookingStatus(UUID userId, UUID performanceId);
-
 	BookingSeatClientResponseDto updateSeatStatus(BookingSeatClientRequestDto request);
 
 	BookingSeatClientResponseDto cancelSeatStatus(BookingSeatClientRequestDto requestDto);
