+++ com.taken_seat.booking-service/src/main/java/com/taken_seat/booking_service/booking/infrastructure/repository/BookingQueryJpaRepository.java
@@ -17,4 +17,6 @@ public interface BookingQueryJpaRepository extends JpaRepository<BookingQuery, U
 	Page<BookingQuery> findAllByUserIdAndDeletedAtIsNull(Pageable pageable, UUID userId);
 
 	Page<BookingQuery> findAllByUserId(Pageable pageable, UUID userId);
+
+	Optional<BookingQuery> findByUserIdAndPerformanceIdAndDeletedAtIsNull(UUID userId, UUID performanceId);
 }
