+++ com.taken_seat.booking-service/src/main/java/com/taken_seat/booking_service/booking/infrastructure/service/BookingClientServiceImpl.java
@@ -5,18 +5,13 @@ import java.util.UUID;
 import org.springframework.stereotype.Service;
 
 import com.taken_seat.booking_service.booking.application.service.BookingClientService;
-import com.taken_seat.booking_service.booking.domain.BookingCommand;
-import com.taken_seat.booking_service.booking.domain.repository.BookingRepository;
 import com.taken_seat.booking_service.booking.presentation.dto.response.SeatLayoutResponseDto;
 import com.taken_seat.booking_service.common.client.PerformanceClient;
 import com.taken_seat.common_service.dto.ApiResponseData;
 import com.taken_seat.common_service.dto.request.BookingSeatClientRequestDto;
 import com.taken_seat.common_service.dto.response.BookingSeatClientResponseDto;
-import com.taken_seat.common_service.dto.response.BookingStatusDto;
 import com.taken_seat.common_service.dto.response.PerformanceStartTimeDto;
 import com.taken_seat.common_service.dto.response.TicketPerformanceClientResponse;
-import com.taken_seat.common_service.exception.customException.BookingException;
-import com.taken_seat.common_service.exception.enums.ResponseCode;
 
 import lombok.RequiredArgsConstructor;
 
@@ -25,15 +20,6 @@ import lombok.RequiredArgsConstructor;
 public class BookingClientServiceImpl implements BookingClientService {
 
 	private final PerformanceClient performanceClient;
-	private final BookingRepository bookingRepository;
-
-	@Override
-	public BookingStatusDto getBookingStatus(UUID userId, UUID performanceId) {
-		BookingCommand bookingCommand = bookingRepository.findByUserIdAndPerformanceId(userId, performanceId)
-			.orElseThrow(() -> new BookingException(ResponseCode.BOOKING_NOT_FOUND_EXCEPTION));
-
-		return new BookingStatusDto(bookingCommand.getBookingStatus().name());
-	}
 
 	@Override
 	public BookingSeatClientResponseDto updateSeatStatus(BookingSeatClientRequestDto request) {
