+++ com.taken_seat.booking-service/src/main/java/com/taken_seat/booking_service/booking/infrastructure/service/BookingCommandService.java
@@ -599,4 +599,5 @@ public class BookingCommandService {
 	private boolean isValidPrice(int price) {
 		return price > 0;
 	}
+
 }
