+++ com.taken_seat.booking-service/src/main/java/com/taken_seat/booking_service/booking/infrastructure/service/BookingServiceImpl.java
@@ -11,12 +11,14 @@ import com.taken_seat.booking_service.booking.application.dto.event.BookingEntit
 import com.taken_seat.booking_service.booking.application.dto.query.BookingAdminListQuery;
 import com.taken_seat.booking_service.booking.application.dto.query.BookingListQuery;
 import com.taken_seat.booking_service.booking.application.dto.query.BookingReadQuery;
+import com.taken_seat.booking_service.booking.application.dto.query.BookingStatusQuery;
 import com.taken_seat.booking_service.booking.application.service.BookingService;
 import com.taken_seat.booking_service.booking.presentation.dto.response.AdminBookingPageResponse;
 import com.taken_seat.booking_service.booking.presentation.dto.response.AdminBookingReadResponse;
 import com.taken_seat.booking_service.booking.presentation.dto.response.BookingCreateResponse;
 import com.taken_seat.booking_service.booking.presentation.dto.response.BookingPageResponse;
 import com.taken_seat.booking_service.booking.presentation.dto.response.BookingReadResponse;
+import com.taken_seat.common_service.dto.response.BookingStatusDto;
 import com.taken_seat.common_service.message.BookingRequestMessage;
 import com.taken_seat.common_service.message.PaymentMessage;
 import com.taken_seat.common_service.message.PaymentRefundMessage;
@@ -99,6 +101,11 @@ public class BookingServiceImpl implements BookingService {
 	@Override
 	public AdminBookingPageResponse adminReadBookings(BookingAdminListQuery query) {
 		return bookingQueryService.adminReadBookings(query);
+	}
+
+	@Override
+	public BookingStatusDto getBookingStatus(BookingStatusQuery query) {
+		return bookingQueryService.getBookingStatus(query);
 	}
 
 	@Override
